// src/alerts/types/Alert.ts
export type AlertStatusChoices = 'NEW' | 'BUSY' | 'DONE';

// Level choices for alerts.
export type AlertLevelChoices = 'CRITICAL' | 'HIGH' | 'MEDIUM' | 'LOW' | 'INFO';

// Outcome choices for alerts.
export type AlertOutcomeChoices = 'completed' | 'duplicate' | 'false positive' | 'N/A' | null;

// Base interface of the alert object from the Cyphon API.
export interface Alert {
  // Unique identifier of the alert.
  id: number;

  // Created date of the data that created the alert.
  content_date: string | null;

  // Date the alert was saved to the database.
  created_date: string;

  // How many times the same alert occurred.
  incidents: number;

  // Priority level of the alert.
  level: AlertLevelChoices;

  // Outcome of the alert.
  outcome: AlertOutcomeChoices;

  // Workflow status of the alert.
  status: AlertStatusChoices;

  // Short description of the alert.
  title: string;

  // Information a user has written about an alert.
  notes: string;

  // Data that created the alert.
  data: Result;

  // User assigned to the alert.
  assigned_user: User | number | null;

  // Distillery the alert came from.
  distillery?: Distillery | DistilleryMinimal | number;

  // User comments associated with the alert.
  comments?: Comment[];

  // Previous actions performed on the alert.
  dispatches?: Dispatch[];
}
// ...
export interface AlertListItem extends Alert {
  // Object of the user assigned to the alert.
  assigned_user: User | null;
  // Object of the distillery the alert came from.
  distillery?: DistilleryMinimal;
}